Downloads/connections-main/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-dotenv
pydantic
passlib[bcrypt]
bcrypt
